Fix a bug where timeout is not honored
There was a bug where the timeout is not honored in multi launch.
The problem is because new rapidfire process will be submitted
if there are other processes running, despite current there is no
active or future jobs in the Fire Server. However, the new rapidfire
should use an updated timeout based on elapsed time, rather than
the original values. To avoid very short launches, the process will
stand down if there is less than 3% of the original timeout avaliable.